Refactor ToolNode implementation and dependencies
- Removed unused import of ToolNode in react.py.
- Added a blank line for readability in the tool __init__.py file.
- Updated test5.py to remove unused import of ValidationError from pydantic.
- Cleaned up imports in test_integration.py for better clarity.
- Commented out an assertion in test_native_tools.py for future reference.
- Removed unused import of append_items in test_utils.py.
- Introduced a new ToolNode package with a modular implementation.
- Created base.py for ToolNode with a focus on modular mixins.
- Added constants.py for ToolNode to avoid circular imports.
- Isolated optional imports in deps.py for ToolNode.
- Implemented various executors for tool providers in executors.py.
- Developed schema utilities for local tool description in schema.py.
- Added shared mixins for logging and event publishing in handler_mixins.py.
