+++ pyagenity/graph/__init__.py
@@ -1,3 +1,16 @@
+"""Graph package public API.
+
+This package exposes the core graph building blocks used by PyAgenity:
+
+- CompiledGraph: Executable, compiled form of a StateGraph
+- Edge: Connection between nodes (optionally conditional)
+- Node: Executable unit wrapping a callable or ToolNode
+- StateGraph: Builder/DSL for constructing graphs
+- ToolNode: Tool registry and executor (modularized under graph.tool_node)
+
+Only these symbols are exported at the package level to keep a clean API surface.
+"""
+
 from .compiled_graph import CompiledGraph
 from .edge import Edge
 from .node import Node
