+++ pyagenity/prebuilt/agent/react.py
@@ -6,7 +6,6 @@ from injectq import InjectQ
 from pyagenity.checkpointer.base_checkpointer import BaseCheckpointer
 from pyagenity.graph.compiled_graph import CompiledGraph
 from pyagenity.graph.state_graph import StateGraph
-from pyagenity.graph.tool_node import ToolNode
 from pyagenity.publisher.base_publisher import BasePublisher
 from pyagenity.state.agent_state import AgentState
 from pyagenity.state.base_context import BaseContextManager
